Conflicts:
	aaaa/src/main/java/aaaa/A.java

+++ aaaa/src/main/java/aaaa/A.java
@@ -2,7 +2,9 @@ package aaaa;
 
 public class A {
 	
-	public void doX(int y){
+
+	public void doX(float y){
+
 		//ha
 	}
